refactor(client): make API address single topics
- choosing a topic is the first thing a user should do
- depending on the type of topic, there may be additional API (e.g. on a
  task topic, a user can create a task subscription)

related to #157

+++ broker-core/src/main/java/org/camunda/tngp/broker/transport/clientapi/CommandResponseWriter.java
@@ -124,7 +124,7 @@ public class CommandResponseWriter implements BufferWriter
 
     protected void reset()
     {
-        topicId = (int) ExecuteCommandResponseEncoder.topicIdNullValue();
+        topicId = ExecuteCommandResponseEncoder.topicIdNullValue();
         longKey = ExecuteCommandResponseEncoder.longKeyNullValue();
         bytesKey.wrap(0, 0);
         eventWriter = null;
